Extract product details and documents components
