+++ apps/store/src/app/[locale]/checkout/components/CartItem/hooks/useOfferDetails.ts
